+++ lib/zammad_api/resources/user.rb
@@ -6,6 +6,8 @@ module ZammadAPI
   module Resources
     class User < Base
       path 'api/v1/users'
+
+      belongs_to :organization, class_name: 'Organization'
     end
   end
 end